Add Project 02
There are eight variables in Program1 called Level, Strength(Str), Dexterity(Dex), Constitution(Con), Intelligence(Int), Wisdom(Wis), Charisma(Cha) and Hitpoints(HP). We must read Level from consol.
The variables; Str, Dex, Con, Int, Wis, and Cha should now be rolled on 3d6. After roll all the stats as one block of six and then display them. Once rolled and displayed the we should have the option to accept or re-roll their stats. Once accepted, we should calculate the bonus for each stat. Each bonus should be 0 at 10, cumulative +1 for each even number above 10 and -1 for each odd number below 10. Hitpoints should be calculated as 1d6+[Con Bonus] per level.
Finally, we must display all variables and their values on the screen and each Bonus should display a + before the number if it is positive, a - before it if it is negative and no sign if the bonus is 0.