Add Program 03
There are eight variables in Program 3 called Level, Strength(Str), Dexterity(Dex), Constitution(Con), Intelligence(Int), Wisdom(Wis), Charisma(Cha) and Hitpoints(HP). We must read Level from consol.
The variables; Str, Dex, Con, Int, Wis, and Cha should now be rolled on 4d6. After roll drop the lowest die and get sum of the three values as a value of the stats.

Once accepted, we should calculate the bonus for each stat. Each bonus should be 0 at 10, cumulative +1 for each even number above 10 and -1 for each odd number below 10. . Hitpoints should be calculated  as 1d6+[Con Bonus] per level. Roll all these stats as one block and then display them. Finally, we must display all variables and their values on the screen and each Bonus should display a + before the number if it is positive, a - before it if it is negative and no sign if the bonus is 0.